fix for skill macro icons
Original code used "explosion" icon for all skill macros, which was never intended to be published. This change uses built-in Foundry icons for Attribute macros and the default macro icon for Skill and Interaction macros. If/when we have our own skill macro icons we can use those instead of the default icon.